main: delete/de-emphasize some legacy code
The manual puzzle instantiation can go away now that we can parse FEN.
The manual testing for `Pawn::move_steps`, et al. doesn't need to run by
default and can be demoted to a never-called function. Now `main` is at
least simple and easy to modify with a different puzzle.